requirements/new.txt
alabaster==0.7.13
amqp==5.1.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
astroid==2.14.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
Babel==2.11.0
backcall==0.2.0
billiard==3.6.4.0
black==23.1.0
boto3==1.26.79
botocore==1.29.79
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coverage==7.2.0
crispy-bootstrap5==0.7
cryptography==39.0.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
Django==4.0.10
django-allauth==0.52.0
django-celery-beat==2.4.0
django-cors-headers==3.14.0
django-coverage-plugin==3.0.0
django-crispy-forms==2.0
django-drf-filepond==0.5.0
django-environ==0.9.0
django-extensions==3.2.1
django-model-utils==4.3.1
django-redis==5.2.0
django-storages==1.13.2
django-stubs==1.15.0
django-stubs-ext==0.7.0
django-timezone-field==5.0
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-stubs==1.9.1
docutils==0.19
drf-spectacular==0.25.1
exceptiongroup==1.1.0
executing==1.2.0
factory-boy==3.2.1
Faker==17.3.0
filelock==3.9.0
flake8==6.0.0
flake8-isort==6.0.0
flower==1.2.0
hiredis==2.2.2
humanize==4.6.0
identify==2.5.18
idna==3.4
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.11
ipython==8.10.0
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
kombu==5.2.4
lazy-object-proxy==1.9.0
livereload==2.6.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==1.0.1
mypy-extensions==1.0.0
nodeenv==1.7.0
oauthlib==3.2.2
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
pluggy==1.0.0
pre-commit==3.1.0
prometheus-client==0.16.0
prompt-toolkit==3.0.37
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==2.6.0
pylint==2.16.2
pylint-celery==0.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyrsistent==0.19.3
pytest==7.2.1
pytest-django==4.5.2
pytest-sugar==0.9.6
python-crontab==2.7.1
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
stack-data==0.6.2
termcolor==2.2.0
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.9.0
types-pytz==2022.7.1.2
types-PyYAML==6.0.12.8
types-requests==2.28.11.15
types-urllib3==1.26.25.8
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
virtualenv==20.19.0
watchdog==2.3.0
watchfiles==0.18.1
wcwidth==0.2.6
Werkzeug==2.2.3
whitenoise==6.4.0
wrapt==1.14.1
